Rename constant and create helpers in runtime/ tests
